// src/pages/main-menu/payment/models/interface.ts
export interface PresentFeeInstData {
  geoHomeZoneList?: string;
  geoHomeZoneCatgList?: string;
  prodAttrList?: string;
  eventInstId?: number;
  acctBookId?: number;
  priceId?: number;
  subBalTypeId?: number;
  effDate?: string;
  expDate?: string;
  presentBal?: number; // default 0
  rateType?: number;
  priceName?: string;
  acctResId?: number;
  acctResName?: string;
  presentSeconds?: number; // default 0
  refAttr?: string;
  isCurrency?: string;
  basicPresentCharge?: number; // default 0
  basicPresentSeconds?: number; // default 0
  currencyPresentCharge?: number; // default 0
  currencyPresentSeconds?: number; // default 0
  spId?: number;
  balanceInfo?: BalanceData;
  seq?: number;
  state?: string;
  stateDate?: string;
  createdDate?: string;
  effSeconds?: number; // default 0
  basicEffSeconds?: number; // default 0
  currencyEffSeconds?: number; // default 0
  cellId?: string;
  presentIntervalUnit?: string;
  presentIntervalTime?: number;
  benefitRevenue?: number;
  balId?: number;
  presentAcctId?: number;
  presentSubsId?: number;
  subBenefitFlag?: string;
  reName?: string;
}

export interface BalanceData {
  balId?: number;
  acctId?: number;
  acctResId?: number;
  grossBal?: number;
  reserveBal?: number;
  consumeBal?: number;
  ratingBal?: number;
  billingBal?: number;
  effDate?: string;
  expDate?: string;
  updateDate?: string;
  preBalance?: number;
  preSuttleBal?: number;
  preExpDate?: string;
  realBal?: number;
  charge?: number;
  seconds?: number;
  ceilLimit?: number;
  floorLimit?: number;
  dailyCeilLimit?: number;
  dailyFloorLimit?: number;
  priority?: number;
  operationType?: string;
  acctResName?: string;
  acctResComments?: string;
  balCode?: number;
  routingId?: number;
  balAcctItemTypeList?: BalAcctItemTypeDto[];
  acctResDto?: AcctResDto;
  subsId?: number;
  effSeconds?: number;
  preEffDate?: string;
  absExpOffset?: number;
  initBal?: number;
  consumeBalCharge?: number;
  custId?: number;
}
export interface BalAcctItemTypeDto {
  balId?: number;
  acctItemTypeId?: number;
  spId?: number;
}

export interface AcctResDto {
  acctResId?: number;
  balType?: number;
  parentAcctResId?: number;
  acctResName?: string;
  isCurrency?: string;
  comments?: string;
  creditLimit?: number;
  remindDay?: number;
  remindValue?: number;
  maxValue?: number;
  refillable?: string;
  paymentForce?: string;
  stdCode?: string;
  isFreeUnit?: string;
  defaultAcctItemTypeId?: number;
  unitTypeId?: number;
  spId?: number;
  unitPrecision?: number;
  ratioMoney?: number;
  ratioPrecision?: number;
  priority?: number;
  extendRule?: string;
  maxAdjustValue?: number;
  maxExpDate?: number;
  maxChgValue?: number;
  resetZero?: string;
  periodClass?: string;
  storeUnit?: number;
  acmThreshold?: number;
  acmType?: string;
  acmUnit?: string;
  acmAmount?: number;
  ceilLimit?: number;
  floorLimit?: number;
  dailyCeilLimit?: number;
  dailyFloorLimit?: number;
  overdraftFlag?: string;
  balanceAggregation?: string;
  rolloverFlag?: string;
  category?: string;
  reservePercentage?: number;
  balCategory?: string;
  gracePeriod?: number;
  maxRollover?: number;
  usageType?: number;
  rewardFlag?: string;
  unlimitedFlag?: string;
  adjustType?: number;
  freeFlag?: string;
  adjustFlag?: string;
  clearFlag?: string;
  clearDays?: number;
  customerFlag?: string;
}